requirements.txt
psycopg==3.2.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
typing_extensions==4.12.2
